feat(FOROME-1051): pr fixes
