+++ src/components/variant-aspects-layout/variant-aspects-layout-grid.tsx
@@ -3,13 +3,15 @@ import GridLayout, { WidthProvider } from 'react-grid-layout'
 import cn from 'classnames'
 
 import { AspectWindow, TWindowToggleHandleParams } from './aspect-window'
-import { IVariantAspectsLayoutGridProps } from './variant-aspects-layout.interface'
 import {
-  adaptLayoutForAspects,
   GRID_LAYOUT_COLS,
   GRID_LAYOUT_CONTAINER_PADDING,
   GRID_LAYOUT_MARGIN,
   GRID_LAYOUT_ROW_HEIGHT,
+} from './variant-aspects-layout.constants'
+import { IVariantAspectsLayoutGridProps } from './variant-aspects-layout.interface'
+import {
+  adaptLayoutForAspects,
   maximizeWindow,
   minimizeWindow,
   useGridHandles,
