+++ src/components/variant-aspects-layout/variant-aspects-layout.constants.ts
@@ -1,0 +1,8 @@
+export const GRID_LAYOUT_COLS = 3
+export const GRID_LAYOUT_ROW_HEIGHT = 40
+export const GRID_LAYOUT_SPACING = 8
+export const GRID_LAYOUT_MARGIN: [number, number] = [
+  GRID_LAYOUT_SPACING,
+  GRID_LAYOUT_SPACING,
+]
+export const GRID_LAYOUT_CONTAINER_PADDING: [number, number] = [16, 16]
